requirements.txt
setuptools~=46.1.3
torch~=1.6.0
torchvision~=0.7.0
albumentations~=0.5.0
numpy~=1.19.1
matplotlib~=3.3.2
Pillow~=7.2.0
